Define the observers as an array of Option for consistent unsubscription
Since removing the observer when an unsubscription happens, the order
of the tracked observers of the subject is modified. This leads
to invalid observers tracking. At some point in the future, an
identifier for each tracked observer could be created, and thus,
a HashSet structure can be implemented. For the moment,
an unsubscription will not modifiy the length of the tracked observers,
but it will be set to None.